Fix the GDJS development watcher making the electron renderer process crash when reloading
Only show in developer changelog
