+++ newIDE/electron-app/app/package.json
@@ -8,6 +8,7 @@
   "homepage": "https://gdevelop.io",
   "main": "main.js",
   "dependencies": {
+    "@electron/remote": "^2.0.8",
     "archiver": "^2.1.1",
     "async": "^2.6.0",
     "axios": "^0.19.1",
@@ -18,10 +19,10 @@
     "electron-is": "^2.4.0",
     "electron-log": "^4.1.1",
     "electron-updater": "4.2.0",
-    "@electron/remote": "^2.0.8",
     "fs-extra": "^3.0.1",
     "jimp": "^0.2.28",
     "live-server": "^1.2.0",
+    "lodash.debounce": "^4.0.8",
     "lodash.throttle": "^4.1.1",
     "minimist": "^1.2.3",
     "ws": "^5.1.1"
